+++ generic/tclParse.c
@@ -2225,7 +2225,7 @@ TclSubstTokens(
 	    if ((appendByteLength == 1) && (utfCharBytes[0] == ' ')
 		    && (tokenPtr->start[1] == '\n')) {
 		if (isLiteral) {
-		    int clPos;
+		    size_t clPos;
 
 		    if (result == 0) {
 			clPos = 0;
